@@ -14,7 +14,7 @@ description: A new Flutter application.
 version: 1.1.0+1
 
 environment:
-  sdk: ">=2.2.0 <3.0.0"
+  sdk: ">=2.6.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -60,6 +60,7 @@ flutter:
   #  - images/a_dot_burr.jpeg
   #  - images/a_dot_ham.jpeg
   assets:
+    - assets/images/
     - locale/i18n_en.json
     - locale/i18n_pl.json
 
