@@ -1,9 +1,9 @@
-import 'package:firebase_database/firebase_database.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:twilight_imperium_tracker/App.dart';
 import 'package:twilight_imperium_tracker/Translations.dart';
 import 'package:twilight_imperium_tracker/feature/game/Game.dart';
+import 'package:twilight_imperium_tracker/feature/game/GameResultExtension.dart';
+import 'package:twilight_imperium_tracker/feature/game/RaceExtension.dart';
 import 'package:twilight_imperium_tracker/feature/game/add_game/add_game_bloc.dart';
 import 'package:twilight_imperium_tracker/feature/game/add_game/add_game_event.dart';
 import 'package:twilight_imperium_tracker/feature/game/add_game/add_game_state.dart';
@@ -27,7 +27,8 @@ class _AddGamePageState extends State<AddGamePage> {
       appBar: AppBar(
         elevation: 8,
         centerTitle: true,
-        title: Text(_translations.text('home_page_title')),),
+        title: Text(_translations.text('home_page_title')),
+      ),
       body: SingleChildScrollView(
         child: Padding(
           padding: const EdgeInsets.all(8.0),
@@ -39,24 +40,31 @@ class _AddGamePageState extends State<AddGamePage> {
                 child: BlocBuilder(
                   bloc: _bloc,
                   builder: (context, state) {
-                    if (state is GameAdded) { return Container(); }
+                    if (state is GameAdded) {
+                      return Container();
+                    }
                     return Column(
                         children: _buildResult(state.game.result, _translations)
-                            ..add(_buildCollectedPoints(state.game.points, _translations))
-                            ..add(_buildGoalPoints(state.game.goal, _translations))
-                            ..addAll(_buildRaceUsed(state.game.raceUsed, _translations))
-                            ..addAll(_buildOpponentsRace(state.game.raceUsed, state.game.opponents, _translations))
-                    );
+                          ..add(_buildCollectedPoints(state.game.points, _translations))
+                          ..add(_buildGoalPoints(state.game.goal, _translations))
+                          ..addAll(_buildRaceUsed(state.game.raceUsed, _translations))
+                          ..addAll(_buildOpponentsRace(state.game.raceUsed, state.game.opponents, _translations)));
                   },
                 ),
-                listener: (context, state) { if (state is GameAdded) { popScreen(context); }},
+                listener: (context, state) {
+                  if (state is GameAdded) {
+                    popScreen(context);
+                  }
+                },
               ),
             ),
           ),
         ),
       ),
       floatingActionButton: FloatingActionButton(
-        onPressed: () { _bloc.add(SaveGame()); },
+        onPressed: () {
+          _bloc.add(SaveGame());
+        },
         tooltip: Translations.of(context).text('add_new'),
         child: Icon(Icons.save),
       ),
@@ -64,41 +72,35 @@ class _AddGamePageState extends State<AddGamePage> {
   }
 
   List<Widget> _buildResult(GameResult result, Translations translations) => [
-    Center(
-        child: Text(
+        Center(
+            child: Text(
           translations.text("pick_result"),
           style: TextStyle(fontSize: 16),
+        )),
+        Row(
+          crossAxisAlignment: CrossAxisAlignment.center,
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: List<Widget>.generate(GameResult.values.length, (int index) {
+            final gameResult = GameResult.values[index];
+            return Wrap(crossAxisAlignment: WrapCrossAlignment.center, children: <Widget>[
+              Radio<GameResult>(
+                  value: GameResult.values[index],
+                  groupValue: result,
+                  onChanged: (GameResult result) {
+                    _bloc.add(GameResultChosen(result: GameResult.values[index]));
+                  }),
+              GestureDetector(
+                child: Text(gameResult.getUserFriendlyResult(translations)),
+                onTap: () {
+                  _bloc.add(GameResultChosen(result: GameResult.values[index]));
+                },
+              )
+            ]);
+          }),
         )
-    ),
-    Row(
-      crossAxisAlignment: CrossAxisAlignment.center,
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: List<Widget>.generate(
-          GameResult.values.length,
-              (int index) {
-            return Wrap(
-                crossAxisAlignment: WrapCrossAlignment.center,
-                children: <Widget>[
-                  Radio<GameResult>(
-                      value: GameResult.values[index],
-                      groupValue: result,
-                      onChanged: (GameResult result) {
-                        _bloc.add(GameResultChosen(result: GameResult.values[index]));
-                      }),
-                  GestureDetector(
-                    child: Text(getUserFriendlyResult(translations, GameResult.values[index])),
-                    onTap: () {
-                      _bloc.add(GameResultChosen(result: GameResult.values[index]));
-                      },
-                  )
-                ]);
-          }
-      ),
-    )
-  ];
+      ];
 
-  Widget _buildCollectedPoints(int points, Translations translations) =>
-      Padding(
+  Widget _buildCollectedPoints(int points, Translations translations) => Padding(
         padding: const EdgeInsets.all(2.0),
         child: Row(
           children: <Widget>[
@@ -115,8 +117,7 @@ class _AddGamePageState extends State<AddGamePage> {
         ),
       );
 
-  Widget _buildGoalPoints(int goal, Translations translations) =>
-      Padding(
+  Widget _buildGoalPoints(int goal, Translations translations) => Padding(
         padding: const EdgeInsets.all(2.0),
         child: Row(
           children: <Widget>[
@@ -134,62 +135,56 @@ class _AddGamePageState extends State<AddGamePage> {
       );
 
   List<Widget> _buildRaceUsed(Race raceUsed, Translations translations) => [
-    Padding(
-      padding: const EdgeInsets.all(8.0),
-      child: Center(
-          child: Text(
+        Padding(
+          padding: const EdgeInsets.all(8.0),
+          child: Center(
+              child: Text(
             translations.text("my_race"),
             style: TextStyle(fontSize: 16),
-          )
-      ),
-    ),
-    Wrap(
-      children: List<Widget>.generate(
-          Race.values.length,
-              (int index) {
+          )),
+        ),
+        Wrap(
+          children: List<Widget>.generate(Race.values.length, (int index) {
+            final race = Race.values[index];
             return Padding(
               padding: const EdgeInsets.all(2.0),
               child: ChoiceChip(
-                label: Text(getUserFriendlyRaceName(translations, Race.values[index])),
+                label: Text(race.getUserFriendlyRaceName(translations)),
                 selected: raceUsed == Race.values[index],
                 onSelected: (bool selected) {
                   _bloc.add(RaceUsedPicked(race: Race.values[index]));
                 },
               ),
             );
-          }
-      ),
-    )
-  ];
+          }),
+        )
+      ];
 
   List<Widget> _buildOpponentsRace(Race raceUsed, List<Race> opponents, Translations translations) => [
-    Padding(
-      padding: const EdgeInsets.all(8.0),
-      child: Center(
-          child: Text(
+        Padding(
+          padding: const EdgeInsets.all(8.0),
+          child: Center(
+              child: Text(
             translations.text("opponents"),
             style: TextStyle(fontSize: 16),
-          )
-      ),
-    ),
-    Wrap(
-      children: List<Widget>.generate(
-          Race.values.length,
-              (int index) {
+          )),
+        ),
+        Wrap(
+          children: List<Widget>.generate(Race.values.length, (int index) {
+            final race = Race.values[index];
             return Padding(
               padding: const EdgeInsets.all(2.0),
               child: ChoiceChip(
-                label: Text(getUserFriendlyRaceName(translations, Race.values[index])),
+                label: Text(race.getUserFriendlyRaceName(translations)),
                 selected: opponents?.contains(Race.values[index]) == true || raceUsed == Race.values[index],
                 onSelected: (bool selected) {
                   _bloc.add(OpponentRaceToggled(race: Race.values[index]));
                 },
               ),
             );
-          }
-      ),
-    )
-  ];
+          }),
+        )
+      ];
 
   @override
   void dispose() {
